wip smooth camera chaser interpolation of ALL PARAMS and view target transition (see description)
Hard to get it right with view blend because when sequences change and the target camera dies, the view target switches to default which is not our chaser. Perhaps I could derive from pawn?

I could:
- register to on destroy of target camera when switching view target
- When a new camera is chosen, unregister

This could still yield problems

How to set smooth chaser camera as default view target?
Where is the code that changes view target when the one we are using is destroyed?